refactor(tehwolfde): group app code into components/embeds/services/showcases
Reorganize the tehwolfde app tree into four intention-revealing folders:
components (app-owned UI), embeds (thin iframe wrappers with a shared
EmbedComponent), services (ThemeService), and showcases (wrappers around the
@tehw0lf/* libraries). Update all routes, imports, and tests accordingly,
add OnPush to NavComponent for consistency, and apply linkcolor to the mobile
nav apps-subheader.
