+++ apps/tehwolfde/src/app/app.component.ts
@@ -1,14 +1,14 @@
 import { NgClass } from '@angular/common';
 import { ChangeDetectionStrategy, Component, inject } from '@angular/core';
 
-import { NavComponent } from './nav/nav.component';
-import { ThemeService } from './theme.service';
+import { NavComponent } from './components/nav/nav.component';
+import { ThemeService } from './services/theme.service';
 
 @Component({
-    selector: 'tehw0lf-root',
-    templateUrl: './app.component.html',
-    imports: [NavComponent, NgClass],
-    changeDetection: ChangeDetectionStrategy.OnPush
+  selector: 'tehw0lf-root',
+  templateUrl: './app.component.html',
+  imports: [NavComponent, NgClass],
+  changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class AppComponent {
   themeService = inject(ThemeService);
