+++ apps/tehwolfde/src/app/components/home/home.component.scss
