+++ apps/tehwolfde/src/app/components/nav/desktop/desktop.component.scss
@@ -1,4 +1,4 @@
-@use '../../../assets/styles/global';
+@use '../../../../assets/styles/global';
 @use '@tehw0lf/mvc';
 @use '../shared-nav';
