+++ apps/tehwolfde/src/app/components/nav/nav.component.ts
@@ -1,4 +1,4 @@
-import { Component } from '@angular/core';
+import { ChangeDetectionStrategy, Component } from '@angular/core';
 
 import { DesktopComponent } from './desktop/desktop.component';
 import { MobileComponent } from './mobile/mobile.component';
@@ -7,6 +7,7 @@ import { MobileComponent } from './mobile/mobile.component';
   selector: 'tehw0lf-nav',
   templateUrl: './nav.component.html',
   styleUrl: './nav.component.scss',
-  imports: [DesktopComponent, MobileComponent]
+  imports: [DesktopComponent, MobileComponent],
+  changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class NavComponent {}
