+++ apps/tehwolfde/src/app/components/nav/shared-nav.scss
@@ -1,4 +1,4 @@
-@use '../../assets/styles/global';
+@use '../../../assets/styles/global';
 
 // Shared styles for both desktop and mobile nav components
