+++ apps/tehwolfde/src/app/embeds/beep/beep.component.ts
