+++ apps/tehwolfde/src/app/embeds/btrain/btrain.component.ts
