+++ apps/tehwolfde/src/app/embeds/numveil/numveil.component.ts
