+++ apps/tehwolfde/src/app/showcases/contact-form/contact-form.component.ts
@@ -2,31 +2,33 @@ import { ChangeDetectionStrategy, Component, computed, inject } from '@angular/c
 import { ContactFormComponent as ContactFormComponent_1 } from '@tehw0lf/contact-form';
 import { of } from 'rxjs';
 
-import { ThemeService } from '../theme.service';
+import { ThemeService } from '../../services/theme.service';
 
 @Component({
-    selector: 'tehw0lf-contact-form',
-    templateUrl: './contact-form.component.html',
-    styleUrls: ['./contact-form.component.scss'],
-    imports: [ContactFormComponent_1],
-    changeDetection: ChangeDetectionStrategy.OnPush
+  selector: 'tehw0lf-contact-form',
+  templateUrl: './contact-form.component.html',
+  styleUrls: ['./contact-form.component.scss'],
+  imports: [ContactFormComponent_1],
+  changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class ContactFormComponent {
   private themeService = inject(ThemeService);
 
   buttonStyle = computed(() => ({
-    'background-color': this.themeService.theme() === 'dark' 
-      ? '#333333' 
-      : 'rgba(255, 255, 255, 0.75)',
+    'background-color':
+      this.themeService.theme() === 'dark'
+        ? '#333333'
+        : 'rgba(255, 255, 255, 0.75)',
     border: 'none',
     color: '#cc7832'
   }));
 
   formStyle = computed(() => ({
     color: '#437da8',
-    'background-color': this.themeService.theme() === 'dark' 
-      ? 'rgba(34, 34, 34, 0.75)' 
-      : 'rgba(255, 255, 255, 0.75)',
+    'background-color':
+      this.themeService.theme() === 'dark'
+        ? 'rgba(34, 34, 34, 0.75)'
+        : 'rgba(255, 255, 255, 0.75)',
     'backdrop-filter': 'blur(50px)',
     'box-shadow': '0 2px 10px rgba(0, 0, 0, 0.075)'
   }));
