+++ apps/tehwolfde/src/app/showcases/wordlist-generator/wordlist-generator.component.ts
@@ -1,0 +1,22 @@
+import { ChangeDetectionStrategy, Component, computed, inject } from '@angular/core';
+import { WordlistGeneratorComponent as WordlistGeneratorComponent_1 } from '@tehw0lf/wordlist-generator';
+
+import { ThemeService } from '../../services/theme.service';
+
+@Component({
+  selector: 'tehw0lf-wordlist-generator',
+  templateUrl: './wordlist-generator.component.html',
+  imports: [WordlistGeneratorComponent_1],
+  changeDetection: ChangeDetectionStrategy.OnPush
+})
+export class WordlistGeneratorComponent {
+  themeService = inject(ThemeService);
+
+  buttonStyle = computed(() => ({
+    'background-color':
+      this.themeService.theme() === 'dark'
+        ? '#333333'
+        : 'rgba(255, 255, 255, 0.75)',
+    color: '#cc7832'
+  }));
+}
