+++ apps/tehwolfde/src/main.ts
@@ -1,12 +1,12 @@
 import { provideHttpClient } from '@angular/common/http';
+import { provideZonelessChangeDetection } from '@angular/core';
 import { bootstrapApplication } from '@angular/platform-browser';
 import { provideAnimations } from '@angular/platform-browser/animations';
 import { provideRouter } from '@angular/router';
-import { provideZonelessChangeDetection } from '@angular/core';
 
 import { AppComponent } from './app/app.component';
 import { routes } from './app/app.routes';
-import { SidenavService } from './app/nav/sidenav.service';
+import { SidenavService } from './app/components/nav/sidenav.service';
 
 bootstrapApplication(AppComponent, {
   providers: [
